pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pie_libs"
version = "0.1.1"
dependencies = [
    "pytz",
    "PyYAML",
    "pymsteams",
    "line-bot-sdk",
]
requires-python = ">=3.8"
authors = [{name = "Pitikorn Khlaisamniang", email = "pitikorn32@gmail.com"}]
maintainers = [{name = "Pitikorn Khlaisamniang", email = "pitikorn32@gmail.com"}]
description = "Python libraries for pitikorn32"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
]

# Tools settings -------------------------------------------------------------------------------------------------------
[tool.setuptools]  # configuration specific to the `setuptools` build backend.
packages = { find = { where = ["."], include = ["pie_libs", "pie_libs.*"] } }